refactor: move OAuthService to RunnerBarCore, rename signIn() → makeSignInURL()
- Add Sources/RunnerBarCore/GitHub/OAuthService.swift (public, no AppKit)
- Update OAuthServiceProtocol: signIn() → makeSignInURL() -> URL?
- Tombstone Sources/RunnerBar/GitHub/OAuthService.swift

NSWorkspace.shared.open(url) is a UI side-effect; it belongs at the
call site (SettingsView / AppDelegate), not inside the OAuth state machine.
With AppKit removed, OAuthService is now independently unit-testable.

Closes #1619, refs #1617
